refactor(pixelflow-core): take the SIMD intrinsics library out of the IR
A crate named IR should not contain a SIMD intrinsics library

`backend/` in pixelflow-ir was two unrelated things wearing one module name:
the `SimdOps` trait family with its per-ISA implementations (3,408 lines of
SSE2/AVX/NEON intrinsics), and the parent module of the JIT emitters. Nothing
connected them — the intrinsics never mention `ExprArena` or `OpKind`, and the
emitters never call `SimdOps`, because compiled kernels emit their own
instructions rather than calling into a Rust trait.

The intrinsics are what `Field` *is*. Their only consumer was pixelflow-core,
reached through an eight-line blanket re-export that existed solely because
they were in the wrong crate. They are the standard library's business, so
they now live in `pixelflow-core/src/backend/`, and the shim is gone.

Two things this makes visible rather than changes. `SimdOps`'s required
methods are the per-ISA primitives and its provided methods are the uniform
expansions built on them — a legal/lowered partition expressed in the trait
system, and the second such partition in the tree after the JIT's. They
disagree: `sin` is required there, expanded here. And this is the layer the
AOT direction intends to eventually *generate* from kernel source rather than
hand-write, which is easier to act on with it sitting beside the stdlib that
would replace it than on loan from the compiler.

With the double duty gone, the emitters stop being nested under a module that
was never about them: `pixelflow_ir::backend::emit` is now `pixelflow_ir::emit`.

`pixelflow-ir` is 18,732 lines, 11,670 of them the emitters. The remaining
~7,000 is the language: the arena, the ops, the passes, the oracle, `Kernel`.
