+++ pixelflow-ir/src/emit/coverage.rs
