+++ pixelflow-ir/src/emit/executable.rs
