+++ pixelflow-ir/src/emit/mod.rs
