+++ pixelflow-ir/tests/spill_pressure.rs
@@ -18,8 +18,8 @@
 
 use pixelflow_ir::OpKind;
 use pixelflow_ir::arena::{ExprArena, ExprId};
-use pixelflow_ir::backend::emit::compile_arena_dag;
 use pixelflow_ir::binding::BindingTable;
+use pixelflow_ir::emit::compile_arena_dag;
 use pixelflow_ir::eval_scalar;
 use pixelflow_ir::jit_manifold::JitManifold;
